controller: modify ConnectionPool.Get() & rename to .GetByNodeID()
This commit modifies ConnectionPool.Get() to take in
driverName and optional nodeID as parameters and now
returns a filtered copy of connections map.
This is done to avoid any deadlock issue that can occur
when unlock function may not be called in previous iteration
of the function.
Unit tests have been added and it now uses assert package.
It is also been renamed to GetByNodeID()

Signed-off-by: Rakshith R <rar@redhat.com>