reclaimspace: detect Kubernetes version for right StagingTargetPath
Kubernetes 1.24 and newer use a different path for staging the volume.
That means the CSI-driver is requested to mount the volume at an other
location, compared to previous versions of Kubernetes. CSI-drivers
implementing the CSI-Addons NodeReclaimSpace procedure, must receive the
correct path, otherwise the driver will not be able to free space and
possibly return an error.

See-also: kubernetes/kubernetes#107065
See-also: https://bugzilla.redhat.com/2096209
Signed-off-by: Niels de Vos <ndevos@redhat.com>