controller: fix csiaddonsnodes object deletion
resolveEndpoint get the pod object to resolve
the ip address as the pod object is deleted
resolveEndpoint will fail to get the pod and
it blocks the csiaddonsnodes object from getting
deleted because of this logs will get flooded
and other csiaddons operation might also fail.

To fix this problem call resolveEndpoint only
if the csiaddonsnodes object is not deleted.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>