replication: reduce the reqeue time for GetReplicationInfo
Reduce the schedule time by half to get the
latest update and also to avoid the
inconsistency between the last sync time in
the VR and the Storage system.

The user can see updates for RPO that are not
stuck in a bad schedule race i.e VR checks
and finds sync time as t-5m and just after
that storage system updates it to t+x.
If we checked every 1/2 of schedule we will
update it to t+x in t+s/2

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>
(cherry picked from commit 8b40a090e65f0f4db373fa88e4e8d76c6d51e960)