NetworkFence: correct check in validating webhook
When a NetworkFence CR is validated, an incorrect comparison is done.
`DeepEqual()` returns `true` when the there are no changes, this is
valid and not an error. Adding `!` to inverse the check corrects the
comparison and only reports an error when the parameters are modified.

See-also: https://bugzilla.redhat.com/2182375
Signed-off-by: Niels de Vos <ndevos@ibm.com>