reclaimspace: add support for namespace annotation
Currently, we only support the reclaimspace
feature on the PVC annotation. However,
this adds support for namespace annotations
as well. If the required PVC annotation is
missing, our controller will check if the
necessary annotation is present on the
namespace. If it is, the scheduling
details will be retrieved from the
namespace annotation and used to create
the ReclaimspaceCronJob.

It's important to
note that the PVC annotation will take
precedence over namespace annotations.

Please be aware that we can only create a
reclaimspaceCronJob if the annotation
exists on the namespace. If an admin wants
to update or remove the annotation on the
namespace, the same action must be taken
on all the PVCs within that namespace.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>