build: add `make docker-generate-protobuf` to update genedated `*.pb.go` files
It seems that not all developers use an updated version of Fedora. For
the contributors that need to update `*.proto` files and re-generate the
related `*.pb.go` files, the new `make docker-generate-protobuf` command
will be very useful.

This includes detection for Podman or Docker in the `Makefile`, so that
developers can use euther (Podman recommended!).

Signed-off-by: Niels de Vos <ndevos@ibm.com>