build: force rebuilding without CGo
Disable building with CGo support, so that only native Golang functions
are used. This makes it possible to build the executables in one
container, and run them in a distroless container.

Fore rebuilding all components with `-a` in the `Makefile`, this
prevents including artifacts that may have been built outside the
contains with CGo enabled.

Signed-off-by: Niels de Vos <ndevos@ibm.com>