ci: disbled unused checks from super-linter
checkov is a linter for infrastructire as code, which we do not use in
this repository. It detects an error under the `vendor/` directory,
which should be skipped either way (filtering not working for this
checker?).

There is no need to run the checked for Go modules. We have other CI
jobs that validate the Go modules already. Because parts of Kubernetes
are vendored (which is large), the Go module checker times out.

Signed-off-by: Niels de Vos <ndevos@ibm.com>