build: update kube-rbac-proxy to v0.18.0
kube-rbac-proxy was pulled from the Google Container Registry, and
Kubernetes managed projects prefer not to use that anymore.

The container-image is (for now, still) maintained outside of the
Kubernetes project, and it is recommended to pull it from quay.io.

While validating the container-image location, it seems that there is a
new version available. Now also using the latest v0.18.0.

Updates: #643
Signed-off-by: Niels de Vos <ndevos@ibm.com>