api: add NetworkFenceClass CRD
adding NetworkFenceClass which is
going to be used to get the client
details to be fenced.

The controller will watch for the NetworkFenceClass
CR and sends a periodic request to all
the sidecars registered with the matching
driver name that registered the
GET_CLIENTS_TO_FENCE caps to get the client
details, once the controller gets the client
details its going to update the CSIAddonsNode
object status with the client information that
can be used to do Network fencing and unfencing.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>