ci: drop linux/arm/v7 from the platforms for building container-images
There is not need for `secrets.BUILD_PLATFORMS`, as the platforms are
not confidential. Instead a variable has been created in the GitHub
repository settings for it now.

The `secrets.BUILD_PLATFORMS` includes linux/arm/v7, which is a slow and
phased out architecture. All reasonable ARM platforms are linux/arm64
now. Dropping the linux/arm/v7 platform makes building the container
images much faster.

Closes: #719
Signed-off-by: Niels de Vos <ndevos@ibm.com>