ci: wait for the controller pod creation before checking status
Sometimes the CI job is running quickly, and the `kubectl wait` for the
controller pod to become ready fails with

    error: no matching resources found

This can happen when the pod does not exist yet. There is a special
`kubectl wait --for=create -l ...` that waits for the pod(s) to be
created. After creation, it is possible to wait for the pod(s) to be
Ready.

See-also: kubernetes/kubectl#1675
Signed-off-by: Niels de Vos <ndevos@ibm.com>