remove pvc finalizers only after VGR is ready to be deleted
finalizers added to PVCs part of VGR should only be removed
after the VGR is ready for deletion, else there can be a race condition
where the VGR is not deleted yet as the group is still undergoing
disabling mirroring and modification operations. Therefore, we
shouldn't remove the finalizers from PVC before the group is ready
for deletion.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>