fix: reclaim jobs are created for pvs in deletion phase
The PersistentVolumeClaimReconciler invoked processReclaimSpace even
when PVCs were already in the deletion phase. This behavior triggered
the creation of reclaim jobs that prevented the underlying volume from
being deleted. Similarly, the ReclaimSpaceCronJobReconciler created
ReclaimSpaceJobs for ReclaimSpaceCronJobs in the deletion phase.  When
managed by a GitOps operator, this resulted in an infinite loop of job
creation and deletion.
The same behavior was also observed for EncryptionKeyRotationCronJobs.
This was also fixed as part of this commit.

Signed-off-by: Ivan Hell <ivan.hell@solunio.com>